* prof_init.c (prof_int32): If long is 4 bytes and int is not, then use long, not int, for prof_int32.
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@16105 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/util/profile/prof_init.c
@@ -18,7 +18,7 @@ typedef short	prof_int32;
 #elif	(SIZEOF_INT == 4)
 typedef int	prof_int32;
 #elif	(SIZEOF_LONG == 4)
-typedef	int	prof_int32;
+typedef long	prof_int32;
 #else	/* SIZEOF_LONG == 4 */
 error(do not have a 4-byte integer type)
 #endif	/* SIZEOF_LONG == 4 */
